nBodyPhaseSpace: split 'nBodyPhaseSpaceGenerator' into two classes
Move those parts of 'nBodyPhaseSpaceGenerator' that calculate weights
and kinematic variables into a new class 'nBodyPhaseSpaceKinematics'.
'nBodyPhaseSpaceGenerator' now inherits from this new class and
generates the intermediate two-body masses and decay angles that are
processed in 'nBodyPhaseSpaceKinematics'. The interface to use
'nBodyPhaseSpaceGenerator' has not changed.
