+++ generators/diffractivePhaseSpace.cc
@@ -48,8 +48,8 @@ diffractivePhaseSpace::diffractivePhaseSpace()
 	  _maxXMassSlices(),
 	  _maxWeightsForXMasses()
 {
-	_phaseSpace.setWeightType    (nBodyPhaseSpaceGenerator::S_U_CHUNG);
-	_phaseSpace.setKinematicsType(nBodyPhaseSpaceGenerator::BLOCK);
+	_phaseSpace.setWeightType    (nBodyPhaseSpaceKinematics::S_U_CHUNG);
+	_phaseSpace.setKinematicsType(nBodyPhaseSpaceKinematics::BLOCK);
 }
