+++ nBodyPhaseSpace/rootPwaGenLinkDef.h
@@ -38,5 +38,6 @@
 #pragma link off all functions;
 
 #pragma link C++ class rpwa::nBodyPhaseSpaceGenerator+;
+#pragma link C++ class rpwa::nBodyPhaseSpaceKinematics+;
 
 #endif
