+++ pyInterface/CMakeLists.txt
@@ -102,6 +102,7 @@ set(SOURCES
 	${HIGHLEVELINTERFACE_SUBDIR}/calcAmplitude_py.cc
 	${HIGHLEVELINTERFACE_SUBDIR}/pwaFit_py.cc
 	${NBODYPHASESPACE_SUBDIR}/nBodyPhaseSpaceGenerator_py.cc
+	${NBODYPHASESPACE_SUBDIR}/nBodyPhaseSpaceKinematics_py.cc
 	${NBODYPHASESPACE_SUBDIR}/randomNumberGenerator_py.cc
 	${PARTIALWAVEFIT_SUBDIR}/complexMatrix_py.cc
 	${PARTIALWAVEFIT_SUBDIR}/fitResult_py.cc
