+++ pyInterface/nBodyPhaseSpace/nBodyPhaseSpaceKinematics_py.h
@@ -1,0 +1,10 @@
+#ifndef NBODYPHASESPACEKINEMATICS_PY_H
+#define NBODYPHASESPACEKINEMATICS_PY_H
+
+namespace rpwa {
+	namespace py {
+		void exportNBodyPhaseSpaceKinematics();
+	}
+}
+
+#endif
